+++ ChangeLog.md
@@ -1,3 +1,7 @@
+### Pending
+* Updates effects on tokens on scene Darkness Level changes.
+* Updates effects on tokens when dropped into scene.
+
 ### Version 0.0.2
 * formatting fixup
 * fixed some issue around clearing effects when moving to Bright lighting condition.